pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "coding-quiz"
version = "0.1.0"
description = ""
authors = ["JAEHO AHN <ajh4234@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = "*"
uvicorn = { version = ">=0.40.0,<0.41.0", extras = ["standard"] }
sqlalchemy = "*"
pydantic = "*"
python-dotenv = "*"
watchdog = "*"
passlib = { version = "*", extras = ["bcrypt"] }
pyjwt = { version = "*", extras = ["crypto"] }
alembic = ">=1.18.3,<2.0.0"
psycopg2-binary = { version = ">=2.9.11,<3.0.0", optional = true }
py-ulid = "^1.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "*"
ruff = "*"

[tool.poetry.extras]
postgres = ["psycopg2-binary"]

[tool.ruff]
line-length = 120
